+++ gs_quant/timeseries/econometrics.py
@@ -383,7 +383,7 @@ def index(x: pd.Series, initial: int = 1) -> pd.Series:
     if not x[i]:
         raise MqValueError('Divide by zero error. Ensure that the first value of series passed to index(...) '
                            'is non-zero')
-    return pd.Series() if i is None else initial * x / x[i]
+    return pd.Series(dtype=float) if i is None else initial * x / x[i]
 
 
 @plot_function
